+++ R/deps_timemachine.R
@@ -1,4 +1,4 @@
-#' Package dependencies for a certain version or time point.
+#' Package dependencies for a certain version or time point
 #' @description Package dependencies from DESCRIPTION files retrieved recursively for certain version or time point.
 #' @param pac character a package name.
 #' @param fields character vector with possible values `c("Depends", "Imports", "LinkingTo", "Suggests")`. Default: `c("Depends", "Imports", "LinkingTo")`
@@ -6,7 +6,7 @@
 #' @param at Date old version of package. Default: NULL
 #' @param recursive logical if to assess the dependencies recursively. Default: TRUE
 #' @note Longer lived version is taken if 2 is available at the same date (switch time).
-#' @return named vector package dependencies and their versions at the release date  of main package plus one day.
+#' @return named vector package dependencies and their versions at the release date of main package plus one day.
 #' @export
 #' @examples
 #' pacs::pac_deps_timemachine("memoise", "0.2.1")
@@ -20,6 +20,7 @@ pac_deps_timemachine <- function(pac,
   stopifnot(all(fields %in% c("Depends", "Imports", "Suggests", "LinkingTo")))
   stopifnot(xor(!is.null(version), !is.null(at)))
   stopifnot(is.logical(recursive))
+  stopifnot(is.null(version) || (length(version) == 1 && is.character(version)))
 
   if (is.null(version)) {
     health <- pac_health(pac, at = at)