+++ R/size.R
@@ -5,11 +5,12 @@
 #' @return numeric size in bytes, to get MB ten divide by `10**6`.
 #' @export
 #' @examples
-#' cat(pacs::pacs_size("stats") / 10**6, "MB")
-pac_size <- function(pac, lib.loc = NULL) {
+#' cat(pacs::pac_size("stats") / 10**6, "MB")
+pac_size <- function(pac,
+                     lib.loc = NULL) {
   stopifnot((length(pac) == 1) && is.character(pac))
   stopifnot(is.null(lib.loc) || all(lib.loc %in% .libPaths()))
-  stopifnot(pac %in% rownames(utils::installed.packages(lib.loc = lib.loc)))
+  stopifnot(pac %in% rownames(installed_packages(lib.loc = lib.loc)))
   found <- try(find.package(pac, lib.loc = lib.loc), silent = TRUE)
   if (inherits(found, "try-error")) {
     0
@@ -18,35 +19,8 @@ pac_size <- function(pac, lib.loc = NULL) {
   }
 }
 
-#' Size of packages.
-#' @description size of packages.
-#' @param pacs character vector packages.
-#' @param lib.loc character vector. Default: NULL
-#' @return numeric vector sizes in bytes, to get MB ten divide by `10**6`.
-#' @export
-#' @examples
-#' cat(pacs::pacs_size("stats") / 10**6, "MB")
-pacs_size <- function(pacs = NULL, lib.loc = NULL) {
-  stopifnot(is.null(pacs) || is.character(pacs))
-  stopifnot(is.null(lib.loc) || all(lib.loc %in% .libPaths()))
-
-  if (!is.null(pacs)) {
-    tocheck <- pacs
-  } else {
-    tocheck <- rownames(utils::installed.packages(lib.loc = lib.loc))
-  }
-
-  dirs <- vapply(
-    tocheck,
-    function(p) pac_size(p, lib.loc = lib.loc),
-    numeric(1)
-  )
-
-  stats::setNames(dirs, tocheck)
-}
-
-#' Package true size
-#' @description Package true size as it takes into account its dependencies.
+#' True size of the package
+#' @description True size of the package as it takes into account its dependencies.
 #' @param pac character a package name.
 #' @param fields character vector, Default: `c("Depends", "Imports", "LinkingTo")`
 #' @param lib.loc character vector, Default: NULL
@@ -57,8 +31,7 @@ pacs_size <- function(pacs = NULL, lib.loc = NULL) {
 #' @examples
 #' # size in MB, with all its dependencies
 #' pacs::pac_true_size("memoise") / 10**6
-#' # exclude packages if at least one other package use it too
-#' pacs::pac_true_size("memoise", exclude_joint = 1L) / 10**6
+#'
 pac_true_size <- function(pac,
                           fields = c("Depends", "Imports", "LinkingTo"),
                           lib.loc = NULL,
@@ -74,5 +47,5 @@ pac_true_size <- function(pac,
     pacs_all <- setdiff(pacs_all, names(Filter(function(x) length(x) > exclude_joint, depsy)))
   }
 
-  sum(pacs_size(unique(c(pac, setdiff(pacs_all, "R"))), lib.loc = lib.loc))
+  sum(vapply(unique(c(pac, setdiff(pacs_all, "R"))), function(x) pac_size(x, lib.loc = lib.loc), numeric(1)))
 }